fix(srv): Change premium increase parameter type to double

+++ 2-insurance-application/src/main/java/de/sri/application/usecases/PolicyManagementImpl.java
@@ -74,7 +74,7 @@ public class PolicyManagementImpl implements PolicyManagement{
 	}
 
 	@Override
-	public void increaseAllPoliciesPremiumBy(int value, int customerId) {
+	public void increaseAllPoliciesPremiumBy(double value, int customerId) {
 		Customer customer = getCustomer(customerId);
 
 		customer.getPolicies().forEach(policy -> {
